Fix time rounding error

+++ robotc/lib/logging.h
@@ -55,7 +55,7 @@ bool log_write(const string tag,const string text)
 	if (!log_setup){
 	    log_init("default.txt",true);
     }
-    float time = (float)(time1[T4]/1000);
+    float time = ((float)time1[T4]/(float)1000);
     string s = "";
     StringFormat(s,"%1.2f:[%s]%s",time,tag,text);
 	WriteText(filehandle, ioresult, s);
